Clear out dead code, duplication, and unjustified assertions
A sweep over the whole source tree, not the recent changes alone. The
codebase was already close: no `any`, no `@ts-ignore`, no `require`, no
TODOs, no commented-out blocks, and `console.log` only in the --help and
--version paths that print before the renderer starts.

What it did turn up:

- `TripleLine` had no callers anywhere. Deleted with its props interface.
- Five internals were exported for nobody: `CODEX_LIMITS_STALE_MS`,
  `GO_LIMITS_STALE_MS`, `SERVER_FUNCTION_KEYS`, `readSpendStore` and
  `defaultUpdateCachePath`. All are module-local now.
- The same ENOENT check was written four times, each casting through
  `NodeJS.ErrnoException`. One `isMissingFile` guard replaces all four.
- `usage-cache` carried its own `record()` that re-implemented `isRecord`
  and asserted the result. It now composes the existing guard.
- Both `decoder.decode(chunk as Uint8Array)` casts were simply unnecessary;
  the stream already types the chunk.
- Two `Object.fromEntries(...) as Record<ProviderId, T>` casts became a
  `byProvider` builder. Adding a fourth provider now fails to compile there
  instead of slipping past an assertion.
- Building the cached rate-limit array with a loop instead of map + some
  lets the compiler see it is non-null, so the last cast goes too.

The only cast left in src is the timer handle in `use-seconds-since`, whose
seam deliberately types the handle as `unknown` so tests can pass a counter.
It now says so.

Five comments in the reducer only restated the case labels under them
("// Refresh lifecycle." above `case "refresh-start"`). Gone. The comments
that record *why* - the per-commit native leak in app.tsx, the signal
handlers in index.tsx - stay, because deleting those is how the reasoning
gets rediscovered the hard way.

Also pins the simplified overview's weekly default with a test. It has been
weekly since the first build and nothing asserted it, so nothing would have
caught it flipping back.

+++ src/data/real/codex-app-server.ts
@@ -292,7 +292,7 @@ async function readDiagnostic(stderr: ReadableStream<Uint8Array>): Promise<strin
   let text = "";
   for await (const chunk of stderr) {
     if (text.length > MAX_DIAGNOSTIC_CHARS) continue;
-    text += decoder.decode(chunk as Uint8Array, { stream: true });
+    text += decoder.decode(chunk, { stream: true });
   }
   const line = text.split("\n").find((candidate) => candidate.trim().length > 0);
   if (!line) return null;
@@ -390,7 +390,7 @@ export async function runRequests(
     const responses = (async () => {
       for await (const chunk of proc.stdout) {
         // stream: true keeps multibyte characters split across chunks intact.
-        pendingText += decoder.decode(chunk as Uint8Array, { stream: true });
+        pendingText += decoder.decode(chunk, { stream: true });
         let newlineIndex = pendingText.indexOf("\n");
         while (newlineIndex >= 0) {
           const line = pendingText.slice(0, newlineIndex).trim();
